clean up some test stuff
