+++ demo/OrderCloud.DemoWebApi/OrderCloud.DemoWebApi.csproj
@@ -9,6 +9,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="AutoFixture" Version="4.15.0" />
     <PackageReference Include="LazyCache" Version="2.1.2" />
     <PackageReference Include="NSubstitute" Version="4.2.2" />
     <PackageReference Include="OrderCloud.SDK" Version="0.10.0" />
